PsychVulkanCore: Implement colorPrecision fallback logic.
In addition to existing types...
0 = 8 bpc unorm RGBA8
1 = 10 bpc unorm RGB10A2
2 = fp16 float RGBA16F
3 = 16 bpc unorm RGBA16

...add new colorPrecision types with auto-fallback. Requests
a high precision, then if that precision is not supported,
falls back to a lower precision, or even a lower precision.

4 = Try 16 bpc RGBA16 unorm ->  10 bpc RGB10A2 unorm.
5 = Try 16 bpc RGBA16 unorm -> fp16 RGBA16F -> 10 bpc RGB10A2 unorm.
6 = Try fp16 RGBA16F -> 10 bpc  RGB10A2 unorm.

The caller will have to probe what precision was ultimately
attained, to adapt its own post-processing accordingly. For that,
PsychVulkanCore('GetHDRProperties', window) now reports back the
field 'ColorPrecision' with the code for the actually used color
precision, ie. 0,1,2,3 for 8/10/fp16/16 bpc.

Note: RGBA16 is not a scanout supported format for swapchains
by any Vulkan driver so far, as far as i know, but added for
completeness and future extensions.