Screen: Suppress some warnings if kPsychNeedFinalizedFBOSinks is used.
The "bad timing in windowed" mode on Linux/X11 gets suppressed if
the flag is set, because the flag means we are using some external
display plugin instead of our OpenGL onscreen window, so it does
not matter if our onscreen window is non-fullscreen. Timing is
determined by whatever that external display backend does.

-> Reduce info message clutter during startup, e.g., if VR HMD's,
   Vulkan/WSI display or HDR display is in use.