PsychtoolboxPostInstallRoutine: Make DrawText text rendering more robust on macOS.
Turns out that if we use GStreamer as the libfontconfig.dylib provider
on macOS then that libfontconfig expects a fonts.conf file with actual
font directory entries for macOS, which the GStreamer provided file does
not do! Workaround this by:

1. Creating our own customized .fonts.conf file with proper path settings
   for macOS, as PsychBasic/macOS_fontconfig.fonts.conf.

2. Let the postinstall routine detect if no user specific $HOME/.fonts.conf
   file exists, and if so, copy our custom file into the proper location.

This allows fontconfig to actually build and verify a proper fontcache
if the fontconfig lib from GStreamer is used. The fontconfig libs provided
by GNU/Octave, HomeBrew, or X11 XQuartz have this set up properly already.

-> Note: I never noticed this, because my build machines also have HomeBrew
   and Octave installed, so this omission never triggered actual failure :/

When we are at it, on non-Linux, let the postinstall routine open a little
invisible PTB onscreen window and draw some dummy text into it, while forcing
use of the plugin text renderer. This way, should a fontconfig cache rebuild
be needed, it will be triggered during post-install, with a proper text message
notifying the user of the process. Maybe this way we can avoid surprise rebuilds
of the cache during work sessions on macOS and Windows, as these often look like
software hangs to the user and trigger false bug reports. We don't need this
trick on Linux, as the OS properly manages fontconfig caches, so a surprise
rebuild is very unlikely to ever happen to a PTB user.