makeftglplugin: Link macOS drawtext plugin against GStreamer fontconfig et al.
In the past we linked against the X11 XQuartz variants of libfreetype and
libfontconfig. This used those if they were installed, with a fallback to
variants from HomeBrew or other system path installed variants.

Turns out this never worked if only GStreamer provided the libraries!
Now we enforce preferred linking against the GStreamer variants, as
use of GStreamer is the recommended approach to get this libs.
Interestingly, if GStreamer is not installed, the linker will still fallback to
system libraries from HomeBrew etc. A catch was that the GStreamer
variants do not come with suitable config files for macOS, so the previous
commit is neccessary for us to provide a suitable .fonts.conf file, so this
will actually work.

So far successfully tested with/without GStreamer, HomeBrew, Octave on
both Octave and Matlab under Mojave and Catalina.

Update 'help DrawTextPlugin' to recommend installation of GStreamer over
installation of XQuartz, as XQuartz is rarely needed for vision science purposes,
but GStreamer is valuable due to adding multi-media support for movie playback.