MOGL: Updated gl_auto.c, first regen since 2014.
This adds over 160 new OpenGL functions - in principle.
Not all of them make sense for PTB user scripts, or are
even usable in a scripting environment, or enabled by
the wrappers. Some of the auto-generated code might need
to be rewritten manually in gl_manual.c ...

But at least we are sort of complete wrt. current OpenGL 4.6
spec and extensions in terms of baseline.

This removes the function glCompileShaderIncludeARB() which,
being in autono.txt since forever, should never have gotten
enabled in the first place and wasn't exposed to user scripts,
as the wrapper was missing.

This also removes the glXXXfOES() functions. They were superseded
in OpenGL-ES by glXXXf() variants with same semantic, and also
never exposed to our users, as no PTB production release ever
had OpenGL-ES support enabled. As such, removing them does not
introduce any functional regressions.

So far this survived a test-compile of moglcore on 64-Bit
Octave + Ubuntu 18.04.5 LTS, and running various OpenGL and
shader based demos didn't show anything amiss, so lets hope
this went well.