Screen: Obey new onscreen window specialflag kPsychExternalDisplayMethod.
This flag if passed to Screen('Openwindow') will tell Screen() that
the onscreen windows OpenGL display surface is not used for actual
visual stimulus presentation. Stimulus drawing and post-processing
is done in the (OpenGL) context(s) of this window, but presentation
to a display is done by some external mechanism/driver/api, e.g.,
the Vulkan/WSI of our Vulkan driver, a VR compositor etc.

Therefore Screen() can skip or suppress certain warnings or checks
related to precise stimulus presentation timing, color reproduction
etc., because the external method of presentation is responsible to
assess such properties and make sure the stimuli are presented at
sufficient quality.

At the moment, this mostly serves to suppress warning and info
messages which apply to OpenGL presentation, but can be misleading
or wrong for external presentation methods. Also, less chatter.

The PsychVulkan() driver sets this flag when Vulkan/WSI presentation
is used for an onscreen window, as the onscreen window just provides
the facilities and OpenGL machinery for stimulus drawing/rendering
and post-processing, but actual presentation(-quality) of final stimuli
is the full responsibility of PsychVulkanCore() and Vulkan/WSI.