Screen('MakeTexture'): Adapt default texture precision in HDR mode.
Change the default setting for the optional 'floatprecision'
parameter in HDR mode: Normally the default is 0 for 8 bpc
fixed point unorm (uint8) precision. This almost never makes
sense in HDR display mode, so change the default to 1 for fp16
RGBA16F textures, which can store colors with the required
minimum precision of fp16 and the extended color range outside
unorm range.

If scripts specifiy 'floatprecision', the request is honored,
but if the parameter is omitted, we use fp16 in HDR mode.

In HDR mode, for 'floatprecision' > 0, ie. actual float, we
also automatically convert passed in uint8 image matrices into
floating point textures - something that does not happen in
non-HDR mode, where this would trigger an error abort.

The conversion converts from uint8 to float, but also applies
a scaling factor, so that the unorm range of uint8 textures is
converted into a meaningful subrange [0; maxsdr], so the texture
fits in if displayed in the HDR onscreen window. maxsdr can be set
via Screen('Hookfunction', win, 'SetHDRScalingFactors', [], maxsdr);
This is done automatically during 'OpenWindow' by the PsychHDR()
driver, according to the unit of color selected, ie. 'Nits' or
'80Nits'. Scaling is chosen, so that uint8 texture pixel values of
255 (the maximum intensity) map to 80 Nits on the HDR display, as
common standards suggest 80 Nits as appropriate reference maximum
value for SDR content on SDR displays.

The applied windowRecord->maxSDRToHDRScaleFactor will also be used
during movie playback of SDR movie content in HDR mode, and the
additional new windowRecord->normalizedToHDRScaleFactor will be
used for playback of HDR movies on HDR displays. This factor is
also set up by PsychHDR() to map a range of 0 - 10000 nits.
Application of scaling factors to movie playback is tbd.

-> Manually tested for compatibility in HDR and SDR mode of float
   and uint8 textures. Seems to work as desired.