Add set of OpenEXR HDR sample images for use in HDR demos.
Psychtoolbox/PsychDemos/OpenEXRImages/ is the folder of choice.
These images are taken from the set of freely available set of
OpenEXR sample images from:

https://github.com/AcademySoftwareFoundation/openexr-images

They are licensed to us and you under a MIT compatible BSD-3 license,
so use and redistribution by us is fine.

We only bundle a small subset of hand picked images, to not bloat
up the download size of Psychtoolbox too much. This adds 17 MB to
the installation size. The mentioned GitHub repo above contains many
more images.



-> License files updated accordingly to give license and credit info.