Screen: Handle SDR movie content in HDR windows for renderable conversion.
8 bpc LDR movie frames in pixelformat 11 need special treatment:

Assign an effective depth of 48 bit (ie. 16 bpc RGB) if the target window is a HDR
window, because then even 8 bpc low dynamic range content must get mapped into the
HDR range for proper display inside the HDR window, ie. the original unorm 0 - 1 range
values must be mapped to 0 - maxSDRToHDRScaleFactor, e.g., up to 80.0 nits. This is not
a problem for direct decode+draw by the planar decode shader during 'DrawTexture', but
in case that immediate conversion into a regular packed pixel color renderable texture
is required via PsychNormalizeTextureOrientation(), or when converting the texture into
an offscreen window for drawing into it, or for Screen('TransformTexture') or similar.
In those cases the LUMINANCE backing texture must get replaced with a backing texture
that is not only colorbuffer renderable, but also capable to contain color values outside
the 0-1 unorm range, iow. we need a floating point texture as backing store. We achieve
this by marking the texture as a high bpc precision texture, so the conversion code in
PsychNormalizeTextureOrientation() will pick a RGBA32F texture as new container, which
can store outside unorm range content with more than 16 bpc linear precision:

=> 8 bit LDR content in SDR window maps to RGBA8 during conversion.
=> 8 bit LDR content in HDR window maps to RGBA32F.
=> 10/12/.. bit LDR and HDR content always maps to RGBA32F.