PsychVulkanCore: Allow export of and waiting for external semaphore.
Optionally, PsychVulkanCore('GetInteropHandle',...) can now create
a interopRenderDoneSemaphore and export/return it to the caller as
a suitable interop handle, if userscript requests this with the new
wantSemaphore=1 flag.

If a semaphore handle is exported, then the driver will wait in each
'Present' operation for that semaphore to become signalled, before
it executes the interop->swapchain blit.

This way an external renderer, e.g., Screen() / OpenGL can signal
render completion on the gpu to mark an interop image ready for
consumption by Vulkan, instead of having to use a blocking cpu
wait via glFinish().

This will be used in followup commits to optimize Screen() interop
performance.