Screen: Allow import and signalling of external render complete semaphores.
Screen('HookFunction', windowPtr, 'ImportDisplayBufferInteropMemory',...)
now optionally takes an external semaphore handle. If the handle is
provided, then it will get imported as a OpenGL semaphore, and this
semaphore will get signalled upon completion of OpenGL rendering into
a finalizedFBO associated with the semaphore.

This allows to signal to external consumers of the finalizedFBO when
it is safe to access the attached interop backing image, ie. when
Screen() is done rendering to the fbo/interop image. This allows to
do all producer->consumer interop synchronization efficiently on the
gpu, instead of requiring a gpu->cpu->gpu roundtrip for host-side
synchronization via glFinish(). Should be more efficient in theory,
and some measurements suggest it is also more efficient in practice.

The semaphore signalling includes some layout transition and signalling
to a target layout for the interop image. This may be more friendly
and forgiving to slightly shaky drivers and avoid some rendering bugs,
but if it really helps is to be seen.

This will be used in a followup commit to reduce handshake overhead
for OpenGL - Vulkan interop between Screen() and PsychVulkanCore()
for higher Vulkan backend performance.