Screen('ReadHDRImage'): Return optional struct 'auxInfo' for exr images.
'auxInfo' is a struct with additional properties and information about an image if
the image file supports such additional meta information. If an image file format
does not support useful additional information then 'auxInfo' will be an empty []
matrix. The fields of the struct are dependent on the file format and specific
image file, so don't assume a fixed set of fields and fieldNames for a returned
'auxInfo' struct.

In case of OpenEXR .exr files, it returns most of the mandatory
attributes of exr images + all optional attributes of type float
or string, plus 'ColorGamut' and 'sampleToNits', the color gamut
definition for the image - red, green, blue primaries and white
point, as well as sample value to absolute nits scaling factor.

If the exr image provides this info in 'sampToNits' it is exposed,
otherwise 0 is returned.

If the exr image provides 'chromaticities' info, then those
primaries and white point are returned. Otherwise the default
gamut for BT-709 is returned, as recommended by the spec.

-> This should allow colorspace conversion from image space to
   onscreen window target colorspace, and unit conversion from
   pixel values into nits for HDR displays.

-> Successfully tested with one exr image that has color gamut
   info, and one exr image that has sampToNits info. Note that
   many (most?) images do not have either of those, so fallback
   defaults apply.