Screen: Zero-Init PsychFBO struct in PsychCreateFBO()!
I did this individually for each variable, and then missed zero-init for coltexid,
instead of simply calloc()ing the memory. This ends badly in some corner-case,
and created a nice hard to track down Heisenbug which took over 12 hours
of maximally tedious work to resolve :(

-> Zero-init coltexid for documentation, but calloc() the whole struct to avoid
this chaos in the future.

-> This caused sporadic failure of HDR content to display on Windows-10 +
     AMD gpu's only, in about 0.25% of all runs. If the bug triggered was
     dependent on coltexid being randomly 1 (potential trigger) or any other
     number (fine), and the specific implementation detail of glGenTextures(),
     which behaves very unforgiving on Win-10 + AMD OpenGL.

-> Bob observes 100% failure rate on a Win-10 system with AMD RX460 gpu.
     I hope this is the same bug and this will fix it...