Screen/Linux: Automatically set 'max bpc' for RandR outputs also under Vulkan.
Factor out the X11 specific code into a helper function
PsychOSEnsureMinimumOutputPrecision(), and also refine
that slightly. For native X11/GLX OpenGL stimulation,
this works as before.

For Vulkan, or any other external display backend provider for
that matter, calling ...
Screen('Hookfunction', win, 'SetWindowBackendOverrides', [], bpc * 3, ...);
... will define framebuffer bit depth as 'bpc' and the function
will also call PsychOSEnsureMinimumOutputPrecision(screen, bpc)
to make sure that all relevant outputs for the onscreen windows
screen will run at an output precision of at least 'bpc'.

What the function actually does is check if more than 8 bpc
are requested for high precision mode, and if so, if 'max bpc'
is greater than 8. If it is already > 8 it does nothing.

Rationale: High precision mode needs highest possible output
precision of the display sink, so if the limit is set to a
meager 8 bpc (startup default for anything but eDP displays),
then raise it to the maximum (e.g., 16 bpc on AMD).

If the limit has been already raised from previous runs of
the PTB script -> nothing to do.

If the user manually chose a 'max bpc' higher than 8 bpc then
we assume that the user knew why they decided to set a new
maximum and we leave it at that - who are we to question our
overlords? This makes sense, e.g., if one connects a HDMI
deep color capable display which claims (EDID) to be able
to handle 12 bpc (this is mandated by HDMI spec for all deep
color displays! Why oh why?), but can actually only do 10 bpc,
simply truncating incoming 12 bpc video to 10 bpc. In such a
case the user is better off manually limiting 'max bpc' to
10 to trigger spatial dithering of > 10 bpc content to 10 bpc,
than sending un-dithered 12 bpc content to a panel that
chops of all but the 10 MSB's.

-> Tested under standard X11 OpenGL and Vulkan with
   10/11/16 bpc OpenGL PTB hacks, and 10/fp16 native
   support of AMD amdvlk.