Screen/Windows: Fix decorationless windows for non-fullscreen/non-GUI mode.
A missing pair of brackets in an if statement made the absence of
kPsychGUIWindow being ignored, and so the WS_POPUP flag for
decorationless windowed windows was not ever assigned.

Weirdly, this worked fine for many years. Some change in the MSVC
compilers treatment since then?

Add some additional brackets to another if statement as well, for
clarity.

Thanks to @suxpert for pointing out the bug and fix.
Fixes #676.