Screen/Linux: Auto-enable dithering to 10 bpc for > 10 bpc fb's on AMD DCE.
For content with more than 10 bpc depths, make sure dithering to
10 bpc is forced on via low-level mmio hacks on AMD DCE display
engines.

Normally the display driver will enable dithering from 12 bpc to
8/10/12 bpc as appropriate, but for displays which run at 10 bpc
signal precision, the amdgpu DisplayCore driver won't enable
dithering to 10 bpc iff the gpu has a DCE display engine.

Code comments suggest this doesn't work on DCE-11, but my testing
shows it works just fine at least on DCE11.2 if force on.

So we add some low level logic which for > 10 bpc framebuffers
detects if outputting to a <= 10 bpc sink, and if so makes sure
that dithering is enabled on DCE display engines.

This problem does not exist on more modern DCN display engines.