PsychHDR: Assign BT-2020 color gamut to window in HDR-10 mode!
This was omitted, so the windows color gamut was not assigned and at
all-zeros default. While that does not matter for rendered/drawn stimuli,
or for movie playback (as movie playback automatically assumes BT-2020
gamut for HDR windows), it meant that calls to ConvertRGBSourceToRGBTargetColorSpace() for converting images silently
turned into no-ops!

Fixed now.