PsychVulkan/PsychVulkanCore: Add workaround for broken AMD Vulkand drivers.
The AMD Vulkan driver for MS-Windows screws up with some gpu models when
a fullscreen window is switched to fullscreen-exclusive mode, as recommended for
best performance and timing robustness.

Symptom is seing a black screen instead of the visual stimulus, and readback of
framebuffer  content via Screen('GetImage') reporting bogus content. This has
been observed on Windows-10 with a AMD Radeon RX 460, with at least driver
version 20.4.2.

Work around this by detecting in PsychVulkan if the AMD gpu under Windows
is a gpu in the badFSEIds list of known bad PCI Id's. If so, then pass in the new
PsychVulkanCore('OpenWindow') flags setting 2, which asks the driver to not
switch to fullscreen-exclusive mode.

This seems to work to work around the problem on the affected system, but it
is unclear if and how much visual timing/timestamping/precision may or may
not be affected by this, or how much general robustness may suffer.

Testing on a RavenRidge Vega 11 by myself and on a RX5600 Navi did not show
this bug. Maybe the driver bug is limited to older pre-Vega gpu's or only some
models?