PsychVulkanCore/Windows: HDR + fp16 or windowed implies scRGB colorspace.
On MS-Windows, for HDR mode with fp16 format, only the scRGB colorspace is
consistently supported across multiple gpu vendors., so use of HDR+fp16 implies
use of the scRGB colorspace.

Additionally for windowed non-fullscreen mode, the Windows DWM must  do
the SDR + HDR compositing and encoding, and that also only supports fp16 +
scRGB color space.

Therefore, whenever we are on MS-Windows and HDR mode 1 is requested,
and fp16 framebuffer or windowed mode is requested, switch to scRGB
colorspace and fp16 format.

This can still be overriden by specifying an explicit colorspace and precision.