Screen/OSX: Deal with more Apple deprecation BS.
Somebody bored and overpaid decided to rename a couple of symbolic
constants, because why not, right? To avoid spill of pointless warnings,
repaint the bikeshed blue. No functional change, because the rename
was literally just to cause half an hour extra work for thousands of ISV's.

Suppress deprecation warnings in PsychCocoaGlue.c from now on.

Note there were more deprecation warnings, which are non-actionable
for us atm.:

NSView shall no longer be used for OpenGL views, but NSOpenGLView,
except its semantics is unsuitable for our specific use case and *drumroll*
NSOpenGLView itself and its docs are deprecated. So we stick to NSView
as long as we can, because life is too short for this BS.

Also disableFlushWindow is deprecated and shall be replaced by a
way more complex and mostly unsuitable mechanism for our purpose,
so no - just say no to drugs. I wish iCompany's employees would do the
same...