DownloadPsychtoolbox: Fix bugs, add auto-install of Subversion via HomeBrew.
Some bugs slipped through due to the peculiarities of my devel machine. Fixed.

Also on macOS, if the svn client is missing for use in the fallback path for
old Matlab and for Octave, try to detect this and give better advise for fixing
to the user.

If HomeBrew is installed, offer to auto-install Subversion from HomeBrew and
then use that, ie. execute "brew install subversion".

Tested on macOS Catalina with both Octave 5.2 and Matlab R2020b.