AssertOpenGL: Add a hack to handle Ubuntu 20.10 libdc1394.25.so
Ubuntu 20.10 ships only with libdc1394.25.so, but our Matlab Screen
mex file for Linux expects libdc1394.22.so. Handle this by detecting
this specific type of Screen "invalid mex file" error due to missing
libdc1394.22.so by detecting if libdc1394.22.so is missing, but
libdc1394.25.so is available. If so, sudo ln -s ... to create a
symlink from libdc1394.22.so -> libdc1394.25.so.

This makes the same Screen.mexa64 Matlab mex file work for both old
Ubuntu 18.04-LTS and 20.04-LTS and the new Ubuntu 20.10.

Also some code indentation cleanup and some help text updates when
we are at it.

Reported by @iandol in issue #187 thanks!