PsychPaidSupportAndServices: Add advertisement for our paid services (I).
PsychPaidSupportAndServices() will describe the community memberships,
priority support and other commercial offerings if called "as is" or
as "help PsychPai...".

If called with a parameter, it will fprintf() a shorter advert which
directs the reader to itself.

Two types of short adverts are printed:

Type 1 is called from routines which typically execute at the end
of installation/update/setup of PTB, e.g., from
PsychtoolboxPostInstallRoutine() - regular download, update, setup,
install, or PsychLinuxConfiguration() - Setup after installation
via NeuroDebian or from Linux distro package manager. This is the
"community membership" - help us and also yourself nag.

Type 2 is called from some error handlers/troubleshooters, especially
if the automated troubleshooter failed to shoot the trouble. It
emphasizes the option of paid priority support.

Also add some section to the main "help Psychtoolbox" help file and
"help PsychDocumentation" top level documentation file.

Still to do is the content of the PsychPaidSupportAndServices main
text, and the updated "Welcome to the Psychtoolbox" splash image.