AssertOpenGL: Add auto-troubleshooter for Octave-5.2 + libdc1394.25.so
Our Screen mex file for Octave 4.4+ (especially Octave 5.2), ie. for
Ubuntu 19.10/20.04-LTS/20.10 and later are built to link against
libdc1394.25.so. For future NeuroDebian or Debian/Ubuntu distro
packages, the maintainers will add proper dependencies, so that
libdc1394.25.so will be present on the system after PTB installation.

However, atm. for downstream, and for our own GitHub distro, and on
Ubuntu 19.10/20.04-LTS, these dependencies are not encoded, so we
may encounter a Ubuntu 20.04-LTS system with only libdc1394.22.so,
not 25.so. Add setup code in AssertOpenGL like for Matlab, which
will detect Octave + 22.so present but 25.so missing, and then offer
to add a suitable symlink to fix the problem.