PsychPaidSupportAndServices(): Allow generation of service authentication tokens.
If a user which has a community membership with priority support,
and therefore is the proud owner of an associated license key,
needs support, then this function will now generate a support
authentication token for them.

The function asks if support is requested, and if the answer is (y)es,
then it asks for the "Order Id" of the community license and the
secret license key.

It combines both with the current system date and time, turns it
into a string, and applies the SHA-256 hash to it. Then it generates
and auth string which combines the public OrderId+date+time with the
SHA-256 hash.

This hash can be used by our dear community member (aka customer)
to request service - ie. as a proof they are deserving of priority
support. The info in the auth token allows us to verify that the
user is in possession of a valid and active license key by:

1. Looking up the license key in our database, using the OrderId
   as lookup key. -> We get his secret licenseKey.

2. Reassembling the string OrderId+date+time+(lookedup in step 1)
   licenseKey.

3. Doing the SHA-256 on our side, and comparing the resulting hash
   with the one provided as part of the authentication token.

4. If the shoe fits then we know we have a valid priority support
   request from a valid and active license key.

Note: We embed the SHA-256 computation routine instead of using our
new PsychComputeSHA(msg, '256') function, so all we need for this
action is self-contained in PsychPaidSupportAndServices.m. This
allows to post the function somewhere else in addition to the
current PTB 3.0.17.