Screen: Apply SetNextCaptureBinSpec option to deviceIndex -1 to -8 as well.
Screen('SetVideoCaptureParameter') option SetNextCaptureBinSpec only applied to devices
with deviceIndex -9. Now we also (ab)use it to
specify videocapture device names and pathes
for deviceIndex -1 to -8, so one can specify
those device specs this way and then use the
'targetmoviefilename' for actually specifying the
target movie name. This way movie recording and
specifications of video capture device names via
name strings are no longer mutually exclusive.

This applies to the 'device', 'device-name' and
'device-path' properties of Windows dshowvideosrc,
ksvideosrc and mfvideosrc, as well as to the 3rd
party 'aravissrc' plugin for capture from GeniCam
video sources.