Screen: Add GStreamer video capture support for mfvideosrc.
mfvideosrc is a new capture plugin on MS-Windows,
which uses the new Microsoft MediaFoundation
framework, which is supposed to replace DirectShow.

mfvideosrc cameras/video sources are now also
enumerated by Screen('VideoCaptureDevices') for
selection via positive deviceIndex 30000-30020, and
via device-name or device-path via deviceIndex -6
and -7 + setting the name as 'targetmoviename' in
'OpenVideoCapture' or 'SetNextCaptureBinSpec='
in 'SetVideoCaptureParameters'. The probing for
mfvideosrc and ksvideorc now uses actual probing
for connected cameras, enumerating only cameras
which exist, instead of always hard-coded exposure
of 5 cameras.

When we are at it, make probing for camera caps
via access to pad's or peer-pads more robust, to
avoid some angry warning messages in some cases.