Screen/GStreamer: Make Videorecording from bins more robust.
Change call flags of gst_parse_bin_from_description_full()
to automatically create a ghost-pad if the bin has exactly
one unlinked internal source or destination pad.

Also if the bin would only contain one element, then
just return that element instead of creating a one-element bin. The required GST_PARSE_FLAG_NO_SINGLE_ELEMENT_BINS is
supported since GStreamer 1.2, so we can safely do this.

This fixes video recording under MS-Windows if one
specifies a deviceIndex -9 type video capture bin via
a gst-launch style string. Untested on other OS'es, but
will hopefully also work there.

Note some peculiarities: This was tested on a MS
Surface Pro 6, whose video cameras can only be properly
operated via the new Mediafoundation framework ie.,
via the new mfvideosrc GStreamer 1.18+ plugin. The
camera - or the plugin - does require specification of
a pixeltype/depth of 6 for YUV-I420 video capture.
Capturing with any other depth or the default type of
3 for RGB8 interleaved does not work and either creates
pixel trash or a crash of Octave! Adding a videoconvert
element to the capture bin did not help at all. It is
unclear so far if this is a problem with our code, or simply
a limitation of the current mfvideosrc plugin, as that
plugin is known to have some limitations compared to
the default ksvideosrc plugin used on Windows.

Possibly Fixes #687. Work time > 10 hours for a total
development cost for this fix of over 1000 Euros.