Screen/Windows: Fix timepoint zero on Windows+GStreamer.
Turns out that as of GStreamer 1.18+ we no longer need
special case handling for setting the clock zero point to
the Screen() startup time. Just as on Linux and macOS,
GStreamer videocapture timestamps on Windows are
now in GetSecs() time, so we do not need to add or
subtract a gs_startupTime offset anymore, ie. the offset
is now a constant zero on all platforms.

-> Testing on Win-10 with GStreamer 1.18.1 confirms
this for both mfvideosrc and ksvideosrc video capture.