VideoCaptureDemo,VideoRecordingDemo: Add usage tip for MS Surface Pro 6.
Mention in the help that pixelFormat 6 YUV-I420 must be explicitely requested for
the builtin cameras of the MS Surface Pro 6 tablet to work. This may apply to other
MS-Surface devices as well, but is only tested on the Surface Pro 6 so far.

With that taken into account, both video capture and recording works as expected
on the tablet.