Screen/Windows: Fix DPI awareness level reporting - Work around Windows bug.
Turns out the GetProcessDpiAwarenessInternal() function that we used on Windows
8.1 and later always worked, but returned the undocumened status code 1 for success,
instead of status code 0 S_OK as it should. This seems to be a Windows bug, present
in at least current Windows-10 20H20 release from October 2020.

Working around this now uses the reporting path for modern Windows-8.1+.

This leads to the pleasant result that Octave-5.2 for Windows is actually reported
as per monitor dpi aware, ie. the best case scenario! Before it was misreported as
only system global dpi aware, which means it would only work on single display
setups or setups where all monitors are either standard 96 dpi or of the same
HiDPI value. It would not work for mixed-DPI multi-monitor setups. This new
per monitor dpi aware result means that PTB with Octave-5.2 on Windows-10
should not get its display timing affected by use on mixed DPI multi-display
setups :).

Note that this behaviour is as i would have expected from reading the QT-5.4+
documentation. Octave uses QT for its GUI, specifically QT-5.14 for Octave 5.2.
QT does all the HiDPI scaling for its apps and defaults to per-monitor scaling
on Windows 8.1 and later. It was always surprising that this default behaviour
did not work on Octave-5+, so now the mystery is solved as a reporting error
due to a Windows API bug.