FitConeFundamentalsWithNomogram: Now works with Octave 6+
The 'optim' Octave-Forge packags since version 1.6.0 supports fmincon(),
and Octave-6.1 now supports use of handles to nested functions. Hurray!

Add code to guide user to optim package use.
Fix some misshaped vlb and vub vectors, which should be each a column vector,
not a row vector. Matlab's fmincon() does not care, although even there it is
technically wrong, but Octave's fmincon() errors out. Fixed.
Remove optimset options 'Diagnostics' and 'LargeScale' which both trigger
"option unsupporte' warnings on Octave, but are both redundant:
'Diagnostics' defaults to 'off' on Matlab, so no need to set it to 'off'.
'LargeScale' does not apply to the selected Algorith 'active-set', so setting it to 'off'
is redundant.

-> With these changes, FitConeFundamentalsTest()  now works on Octave-6.1.
-> Fixes #137.