LjgToXYZ(): Make it work with Octave-6 and later.
Convert to use of a nested function handle for the target function
to minimize. LjgToXYZFun.m this way becomes redundant as a
standalone M file.

This should allow passing extra static parameters as variables
from the surrounding parent function of the nested function,
thereby circumventing the lack of Octave's fmincon() function
ability to pass in optional 11th, 12th, ... parameters. Note
that this ability although apparently present in Matlab is
not documented by R2020b docs!