Add PsychContributed/Eduloggers: Simple Edulogger support.
Contributed by Todd Parsons (@TEParsons) under MIT license,
thanks! This is unmaintained and untested code, as it was
partially rewritten to reach the minimal quality standard
of PTB by Mario Kleiner, but could not be tested after the
rewrite due to lack of Edulogger hardware. Hence the location
in the PsychContributed folder for the time being, instead of
proper placement in the PsychHardware folder.

Closes #526.