PsychVulkan[Core]: Add VK_GOOGLE_DISPLAY_TIMING support.
If VK_GOOGLE_DISPLAY_TIMING is supported by Vulkan driver then use it
for scheduling stimulus onset time and for retrieving native high
reliability high precision stimulus onset timestamps. No trickery needed
via our Screen() hacks. Otherwise fall back to our trickery.

Tested with Keith Packards wip implementation of this extension for
Linux + Mesa Vulkan OSS drivers. Specifically lightly tested with the
build for radv in Mesa 21.0.0-devel as of 12.12.2020. See for reference:

https://gitlab.freedesktop.org/keithp/mesa/-/tree/present-timing

One weirdness we currently work around is that for the first present,
we do not get a timestamp back. In this case we time out after 100 msecs
and let it slip if it was the first present in a session, otherwise
we make warning noise. Todo: Find out if this is expected/valid behaviour
according to spec, or an implementation bug/flaw.

-> Not yet tested with timing equipment, but results look reasonable
   and agree with results from our timestamping hack which was tested
   with timing equipment.

-> The wip EXT_present_timing extension is not yet enabled for radv,
   maybe only for anvil? The extension spec is not finalized yet, so
   this is of low relevance anyway at this point.