Fixes for Python package setup on Windows.
After the package restructuring, the portaudio and libusb DLLs weren't copied
to the correct location. This let to ImportErrors ("DLL load failed") after installation.

This also adds numpy to 'install_requires', so that it will automatically be installed
if it is not already.

This also bumps the Python package version to 3.0.17.5, to match the latest tag.