OffsetRect(): Allow single rect as both row and column vector.
Instead of only row vector. Fixes failure of, e.g.,
ProceduralGarboriumDemo(1) if nrects == 1.

However, we still fail ProceduralGarboriumDemo(4),
because a 4 rects, the passed in matrix is 4x4 and
so there's ambiguity if this is 4 rects, one in each
row, or 4 rects, one in each column. Combined with
ambiguity with other parameters we end in a nice little
mess of poorly defined behaviour.

=> It was a mistake to ever allow so many different behaviours
   of some of the PsychRect functions, with the behaviour
   switching based on the shape of the input (rows or cols,
   matrix or single vector, etc.). We should have created
   dedicated functions, or require an explicit parameter that
   specified wanted behaviour. This guessing game is too much
   filled with ambiguity and suprises. Bad api design...

   We can't easily go back without breaking backwards compat,
   but let this be a lesson for future similar situations.

Reported by @iandol.