PsychtoolboxPostInstallRoutine: Treat Octave on ARM always as Octave 3.8.
Currently we can have one common set of 32-Bit mex files for all
versions of 32-Bit Octave on Linux for 32-Bit ARMv7 architecture,
e.g., for Raspbian / the RaspberryPi 0/1/2/3/4. We store these in
the old Octave3LinuxARM/ folder, so make sure our setup routine
actually sets that path regardless of Octave version.

PTB for Octave+ARM is currently tested only on 32-Bit Raspbian/
RaspberryPi OS on the RPi 2B and 4 (actually RaspberryPi 400 Kit),
with whatever the Octave version on current Raspbian is.

The expectation is that it should work on Octave 3.8 - 5.2, but
testing atm. is against Octave 4.4.1.