Add new "help RaspberryPiSetup" with docs about use of the RPi.
The RPi-4 is in rather excellent shape, as further testing
confirmed. The only known serious limitation atm. is that
audio playback over HDMI doesn't work with PsychPortAudio.
That already seems to have been the case with the RPi-2 though,
and as a workaround, the onboard headphone audio output will
work, as will Audio playback via GStreamer movie playback, or
use of an external UAC USB audio card/dongle.

This concludes initial bringup of Psychtoolbox for the RPi 4/400,
as well as initial testing. Lets see if anybody cares enough to
fund further work.