KbCheck: Add workaround for missing ESC key on dead touchbar.
When running a Linux kernel without Apple touchbar driver on
a touchbar MacBookPro, e.g., a development or not-yet-released
rc kernel, where the touchbar dkms support isn't available yet,
we won't have the touchbar and therefore won't have the ESCape
key, which is crucial for many test scripts.

Add a workaround to KbCheck: Map the ~ key left to the 1! key
also to ESCAPE.