Replace all internal calls to range() by calls to new psychrange().
psychrange() is a drop-in replacement for range() if range() is missing,
e.g., on a Matlab without statistics toolbox installed. As most other
stats functions that we need internally are available in core Matlab
and Octave, this allows us to do without statistics toolbox for simple
use cases. Useful to shave off a GB of Matlab install size on smaller
test machines.