MultiTouchMinimalDemo:  Add a small potential performance optimization.
We currently use the touch blob 'id' directly to index into the blobcol cell array.
This means that for a large 'id', e.g., because the id's are montonically increasing
with each touch-point sequence, and the system is running already for quite a
whil, that we may have hundreds or thousands of empty entries in blobcol at
the beginning. Our processing for-loops for drawing will have to iterater over all
these slots just to then ignore them. Try to save some processing overhead by
adjusting the 'id' field to always start at 1 for a session. A better solution would
be to design this differently, but as a quick test, it may help performance in some
cases.