Windows: Require WINVER >= 0x601 (Windows-7) at a minimum.
To allow use of Windows-7 and later only api.
This is only needed for building on GNU/Octave-6.1 atm., which defaults to
a lower WINVER for Windows-XP iirc. When building on MSVC, the Win-10
SDK will already default this to an appropriate value, or at least so i think
until testing proves me wrong.