Windows/Base/PsychPlatform.h: Define WINVER for Octave build manually.
We need WINVER >= 0x0601 for Windows-7 SDK functionality for some future
work. The Windows-10 SDK from MSVC 2019 defines it already to a higher value,
but the Octave-6.1 MinGW64 SDK doesn't, so define WINVER et al to 0x0601 on
Octave builds, to get at least new functions from Windows-7.

This will likely put a final end to running on Windows Vista or earlier, but as
we officially only  support Windows-10 anyway, this is fine.