ScriptingGlue/Matlab: PsychCopyInPointerArg() always accepts uint32 and uint64.
Before it expected uint32 on 32-Bit architectures, uint64 on 64-Bit archs.
Now it accepts both uint32 and uint64 regardless of arch, but then casts
to a suitable size void* for the arch.

Errors out if on a 32-Bit architecture, a uint64 is passed in, whose actual
value exceeds the representable range of 32-Bit pointers, iow. truncation
and mayhem would happen.

Why? Because we sometimes (ab)use PsychCopyInPointerArg to just get
32-Bit or 64-Bit uint's without precision loss, so a system might get confronted
with a "wrong" size for the architecture and has to deal with it.

In practice this would not have been a problem on 64-Bit systems, as all the
quantities we pass so far are either pointers or uint64 (PsychVulkan). On a
32-Bit system, PsychVulkan would have failed though, as it would not have
accepted a uint64. Given the lack of testing on 32-Bit by now, except for the
RaspberryPi where we don't support Vulkan yet, i didn't notice.