Windows/Base/PsychPlatform.h: Bump WINVER for Octave to 0x0602 manually.
Lets push this further and require 0x0602 from now on. This enables functions
added in Windows 8.0 / Win 8 SDK, and obviously at least Windows-8 to use
these functions. First use case will be to get access to the CreateWindowEx constant
WS_EX_NOREDIRECTIONBITMAP introduced in Windows 8.

On MSVC 2019, we already use the Win-10 SDK anyway, so this is only needed for
Octave-6.1.