Screen/Windows: Add new visual diagnostic for DWM compositor interference.
If Screen('Preference', 'VisualDebugLevel', 6) is set and we runon Windows-8 or
later, and a fullscreen window is opened - which shouldn't be subject to DWM
composition - then create the onscreen window with DWM redirection surface
disabled. This will cause any DWM composited windows client area to not
display at all, but instead just show transparent.

This is a clear visual indicator that a window is subject to unwanted DWM
composition, iow. if  the DWM is the root cause for visual timing and
timestamping problems.

This is meant as yet another diagnostic aid for Windows-10 timing problems.
We require Win8-SDK or later, as enabled in the previous commit via  WINVER
0x0602.