GetMouseIndices: Only expose "master pointer" on Windows.
On Windows we currently only have one unified mouse-device,
ie. more or less what a 'masterPointer' would be on Linux/X11
with a default setup (= Only one corepointer). So ignore the
'typeOnly' parameter on Windows and always just return the
one master pointer.

Removes redundant enumeration of the one and only mouse,
showing the same handle and device twice.