MinimumMotionExp: Adapt for triple-buffered OpenGL implementations.
Setups with more than double-buffering would flicker with the old
code. Let's not use dontclear=2 but only dontclear=1 and enable use
of the virtual framebuffer for efficient implementation of dontclear=1.
This makes the animation loop in CRS Bits+ / VPixx L48 mode free of
flicker again.

Also some other cleanups, when we are here anyway.