Add Denis Pelli's MeasureLuminancePrecision.m as baseline to PsychAlpha/
This is a nice script in principle, with a nice automated fitting
procedure, to automatically assess the visual luminance precision
for given Psychtoolbox display modes on given hardware.

It presents a range of luminance intensity levels, measures the
displays response with a ColorCal-II, then fits a model to the
measured response curve to estimate the bit depth of luminance
reproduction, e.g., 8 bpc, 10 bpc, 11 bpc, 12 bpc ...

The script has proven useful quite often, but always needed to be
hacked up for each configuration, and also to make it compatible
with Octave.

It is in strong need of a major cleanup and streamlining to make
it more generally useful without need for hacking, and to demonstrate
best practices. We put it into PsychAlpha for the time being as a
starting point for the cleanup.