Screen: Add new return field to 'GetWindowInfo' for gfx-api interop.
Expose the windows targetSpecific.deviceContext as new return field
SysWindowInteropHandle, a 64-Bit pointer/number. This can be useful
for graphics api interop. Currently used on macOS to expose a pointer
to the CALayer backing Cocoa NSWindow's if those are used to implement
onscreen windows. If CGL fullscreen windows are used instead, then this
parameter will be meaningless. The CALayer iff layer backing is used, will
be of type CAMetalLayer, ie. renderable and displayable by the macOS
Metal graphics api and windowing system interface. This is needed for
use of our windows by Vulkan / MoltenVK ICD / Metal.