Screen/OSX: Add support for kPsychExternalDisplayMethod via Metal.
If kPsychExternalDisplayMethod is requested for an onscreen window,
then we will always create and use a NSWindow via Cocoa, instead of
a direct display CGL setup. For this window's NSView contentView, we
need to use a CoreAnimation backing layer as window framebuffer.
Specifically we need to create and attach a CAMetalLayer as backing
layer.

This happens after Screen('OpenWindow') has opened and displayed
the window in standard OpenGL mode, initialized OpenGL, displayed the
welcome splash screen via OpenGL and ran all the timing startup tests
and calibrations, as we need the results of those calibrations later on.

Shortly before 'OpenWindow' returns control to the calling M-File, we
attach our CAMetalLayer. This effectively prevents future OpenGL
rendering and display for our NSWindow, but allows Metal to render into
or display from the NSWindow.

Why Metal? Because we use this method to use Vulkan as display backend,
and the only existing Vulkan ICD for macOS - MoltenVK - implements
Vulkan on top of Apples Metal rendering api, so it requires a CAMetalLayer
as rendering target for onscreen display via Vulkan swapChains.

Some observation so far:

- Windowed display and fullscreen display via a NSWindow that is a tad
transparent, and thereby can not bypass the Quartz compositor, works at
ok framerate and without artifacts.

- Fullscreen decorationless topmost opaque windows - which may allow for
compositor bypass - show lower update rates (!) and visible tearing/rendering
artifacts! I don't know if this is a Metal/macOS bug, or an interaction bug
between MoltenVK and Metal, but it is a potential source of trouble. This
as tested on 10.15.7 on AMD Radeon Pro 560, MBP 2017 15 inch on the
internal Retina panel, with MoltenVK 1.1.1 (released 9th December 2020).

-> Baseline implementation to get anything Vulkan attached to our
onscreen window for Vulkan/WSI presentation.