PsychVulkanCore: Turn pastTiming single record into a 1000 Element array.
Fetch up to 1000 present history records at once, instead of only the oldest
one (FIFO order fetch). Then select the one with matching presentID for a
given present operation.

This is needed to work around bugs in the macOS MoltenVK Vulkan driver,
on a high quality Linux Vulkan implementation, we will only get 1 record per
present with both the old and this new multi-fetch implementation.

So it will help on broken macOS and should not hurt on well working Linux.