PsychVulkanCore: Fix basic timing on macOS + MoltenVK 1.1.1
At least as of the following MoltenVK version...

DriverInfo: '1.1.1'
DriverVersion: '0.2.1909'
DriverVersionRaw: 10101

...MoltenVK's VK_GOOGLE_display_timing extension has a bug:
The queue with present history does not empty if one fetches records
from it. Iow. fetching does not dequeue the oldest (FIFO) records,
shrinking the queue accordingly, but just "peeks" at the data, keeping
the queue size as is. Therefore the queue can only grow, up to a maximum
of 60 entries atm., but never shrink.

This means that repeated fetches will return the same records repeatedly,
so waiting after a present for a queue empty -> non-empty transition does
not work to poll for present completion. Instead we need to fetch all records
available in the queue, try to find the one with the presentId of our present
request, and use that one for further processing. If the queue does not
contain our presentId for the most recent present request, then that means
that the present has not yet completed -> We need to repeat the full query
procedure, after sleeping for 0.1 msecs until success or tQueryTimeout of
0.5 seconds is reached, which suggests something really went wrong.

As this is not how Vulkan should behave, we only enable this hack/workaround
for macOS. On other OS'es this would indicate some serious failure and we
therefore simply abort the present operation.

This makes basic present scheduling and timestamping work ok'ish. I still
need to verify if the reported timestamps have anything to do with reality,
iow. do the usual hardware timing measurement procedure. But that needs
first work on basic OpenGL-Vulkan interop on macOS.

Limitations:

- The first reported present (presentId 0) always reports a timestamp of 0.
Why that is, i don't know, but it comes from the Metal layer, so basically it
is a macOS Metal bug.

- If we don't reach direct-to-display mode / compositor bypass, then we
have sporadic timestamping failure with 0 timestamps --> macOS Metal bug.

- Very occassionally we see a lack of a returned present record, most often
for presentId 1, almost certainly due to another macOS bug. Luckily we can
ignore such failures for the first 4 presents, as they happen during the startup
sequence of PsychImaging('OpenWindow'), so won't affect actual users scripts.

- Most severe: I can't get more than 30 fps (one present every 33.333 msecs)
on the 60 Hz Retina panel of the MBP 2017. Not yet investigated if this is a
performance problem which would go away on. a non-Retina or lower resolution
display, or faster machine. If this would be an inherent problem / design flaw
in Apples timestamping logic or scheduling for Metal presents, we'd be quite
screwed -- nothing we could do to get better performance and proper timing!

-> This hack is hopefully temporary. A proper solution will be to fix MoltenVK
to be spec-compliant and get my fix integrated in a future MoltenVK release.
Then we could drop this hack again for a more simple and maintainable
implementation.

-> Also needs retesting on Linux, to make sure we didn't accidentally break
the properly working system for the benefit of the trainwreck.