PsychVulkanCore: Add OpenGL-Vulkan interop for macOS / MoltenVK.
As Vulkan for macOS does not yet have any official specification of how
interop is supposed to work, we have to use a different approach than the
one taken on Linux/Window, where the interop images backing VKMemory
is exported from Vulkan and imported into OpenGL. Also Apples OpenGL
does not support GL_memory_objects or semaphores.

Instead we directly link to MoltenVK and compile PsychVulkanCore.c as
Objective-C, so we can use a MoltenVK proprietary method of image
sharing:

1. Use vkUseIOSurfaceMVK() to ask MVK to allocate the interop VKImage
with a macOS IOSurface as backing store.

2. Query the IOSurface via vkGetIOSurfaceMVK().

3. Use macOS CGL method CGLTexImageIOSurface2D() to create a
GL_TEXTURE_RECTANGLE texture from the IOSurface on Screen's
main OpenGL context for the onscreen window.

4. Export that OpenGL texture handle to Screen, so it can be imported
and used as color attachment for the finalizedFBO[0].

5. Render to that texture as usual -> Render into IOSurface, which backs
the MTLTexture Metal texture which backs the VKImage interop image.

This is the same principle as on Linux/Windows, but using IOSurface's
instead of exported VKDeviceMemory.

Note that the IOSurface is created by MoltenVK as shared IOSurface.

Another approach supported by MVK would be to use OpenGL + Metal
texture sharing. This is more complicated, and supposedly not more
efficient, but would be an alternative if this proves to not work well. One
would create a CVPIxelBuffer which is compatible with OpenGL and Metal,
then create OpenGL and Metal textures backed by that CVPixelBuffer,
then assign the Metal texture as VKImage backing texture, via a MVK
proprietary function - which requires to use Objective-C (gaah!), and
use the OpenGL texture on Screen's side.

We first try the IOSurface approach as it allows us to stay away from
annoying ObjC and have more simple code and logic.

Linking directly to libMoltenVK.dylib is important here, as the standard
Vulkan loader and layers do not support the MVK proprietary interop
support functions and would cause "undefined behaviour" aka crashes!

The vkUseIOSurfaceMVK and vkGetIOSurfaceMVK function prototypes
are only exposed to ObjC code for unclear reasons, maybe an omission.
Therefore we compile as ObjC on Matlab. For Octave, where this is a bit
more hassle, we instead define the relevant function prototypes ourselves
if !defined __OBJC__, to sidestep the need to compile other than C.