PsychImaging/PsychVulkan: Enable OpenGL-Vulkan interop on macOS.
This requires minor modifications:

We use Screen('Hookfunction', win, 'SetDisplayBufferTextures',...) on macOS
instead of Screen('Hookfunction', win, 'ImportDisplayBufferInteropMemory',...),
as we import a ready made OpenGL texture (backed by an IOSurface exported
from MoltenVK for the interop VKImage), instead of Vulkan device memory.

This requires the kPsychUseExternalSinkTextures imagingMode flag, so that
Screen()'s sets up the imaging pipeline and finalizedFBO's for use of an
externally created and injected OpenGL texture. -> Modify PsychImaging.m

PsychVulkan also needs to *not* override ovrfbOverrideRect, as it does on
other OS'es, so we can deal with Retina displays and their fake fb size.

As we do not have semaphores on prehistoric Apple OpenGL, we use the
glFinish() method for sync between OpenGL and Vulkan present.

As we don't have device UUID's on prehistoric Apple OpenGL, we always
use the first logical Vulkan gpu (gpuIndex 1) for interop. Luckily Apple
machines are single-gpu or the dual-gpu laptops will use the high-perf
gpu under OpenGL with Psychtoolbox, which seems to match gpuIndex 1
on the Vulkan side. A multi-gpu MacPro might have trouble with this
approach, but one can't please everybody.

So how well does this work? Results are suboptimal:

OpenGL interop seems to work in the sense that we get the rendered
images to display properly. However, every 4th Presented image seems
to not contain the rendered image, an image that is about 3 frames in the
past! This is probably not actually a problem related to interop, but a
symptom that we can only see now that we can render and display other
images than black-white alternating pairs. It looks as if the swapchain
wouldn't have the 2 images it claims to have for double-buffering, but
as if it had at least 3 images for triple-buffering, and something went wrong
with presentation order in a. weird way. Needs more investigation, as
currently the results are less than satisfying.

-> Backup commit, the basics work, and this establishes a new baseline
for further testing.